Quite a few changes uncluding redoing the database and backend

+++ backend/backend/Repositories/StoryPartRepository.cs
@@ -1,8 +1,8 @@
-﻿using Microsoft.EntityFrameworkCore;
-using backend.Data;
-using backend.Models;
+﻿using backend.Models;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
+using backend.Data;
 
 namespace backend.Repositories
 {
@@ -17,12 +17,16 @@ namespace backend.Repositories
 
         public async Task<IEnumerable<StoryPart>> GetStoryPartsAsync()
         {
-            return await _context.StoryParts.ToListAsync();
+            return await _context.StoryParts
+                .Include(sp => sp.Story) // Include related entities if needed
+                .ToListAsync();
         }
 
         public async Task<StoryPart> GetStoryPartByIdAsync(int id)
         {
-            return await _context.StoryParts.FindAsync(id);
+            return await _context.StoryParts
+                .Include(sp => sp.Story) // Include related entities if needed
+                .FirstOrDefaultAsync(sp => sp.PartId == id);
         }
 
         public async Task<StoryPart> AddStoryPartAsync(StoryPart storyPart)
@@ -32,18 +36,20 @@ namespace backend.Repositories
             return storyPart;
         }
 
-        public async Task<StoryPart> UpdateStoryPartAsync(StoryPart storyPart)
+        public async Task UpdateStoryPartAsync(StoryPart storyPart)
         {
             _context.Entry(storyPart).State = EntityState.Modified;
             await _context.SaveChangesAsync();
-            return storyPart;
         }
 
         public async Task DeleteStoryPartAsync(int id)
         {
-            var storyPart = await _context.StoryParts.FindAsync(id);
-            _context.StoryParts.Remove(storyPart);
-            await _context.SaveChangesAsync();
+            var storyPart = await GetStoryPartByIdAsync(id);
+            if (storyPart != null)
+            {
+                _context.StoryParts.Remove(storyPart);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
